Move reducer.js to src/
